feat: fetch featured posts from Supabase to fix homepage links
Ensure homepage links match actual blog posts in Supabase.

+++ scripts/create-view-increment-function.sql
@@ -1,5 +1,5 @@
 -- Create a function to safely increment post views
-CREATE OR REPLACE FUNCTION increment_post_views(post_id BIGINT)
+CREATE OR REPLACE FUNCTION increment_post_views(post_id INT)
 RETURNS VOID AS $$
 BEGIN
   UPDATE blog_posts
